+++ apps/ungdomsytelse-deltaker/src/apps/innsyn/components/oppgavebekreftelse/parts/UløstOppgavebekreftelse.tsx
@@ -1,14 +1,17 @@
 import { Alert, Box, GuidePanel, Heading, VStack } from '@navikt/ds-react';
 import { useEffect, useRef, useState } from 'react';
-import UtalelseForm from '../../../forms/uttalelse-form/UtalelseForm';
-import ForsideLenkeButton from '../../../atoms/forside-lenke-button/ForsideLenkeButton';
-import { usePrevious } from '@navikt/sif-common-hooks';
 import { useNavigate } from 'react-router-dom';
+import { UngdomsytelseOppgaveUttalelseDto } from '@navikt/k9-brukerdialog-prosessering-api';
+import { usePrevious } from '@navikt/sif-common-hooks';
 import { BekreftelseOppgave } from '@navikt/ung-common';
-import { OppgavebekreftelseTekster } from '../Oppgavebekreftelse';
-import OppgaveStatusTag from '../../oppgave-status-tag/OppgaveStatusTag';
-import { getOppgaveStatusText } from '../../../utils/textUtils';
+import { useAnalyticsInstance } from '../../../../../utils/analytics';
 import { AppRoutes } from '../../../../../utils/AppRoutes';
+import ForsideLenkeButton from '../../../atoms/forside-lenke-button/ForsideLenkeButton';
+import UtalelseForm from '../../../forms/uttalelse-form/UtalelseForm';
+import { LogMetaInfoType, logUtils } from '../../../utils/logUtils';
+import { getOppgaveStatusText } from '../../../utils/textUtils';
+import OppgaveStatusTag from '../../oppgave-status-tag/OppgaveStatusTag';
+import { OppgavebekreftelseTekster } from '../Oppgavebekreftelse';
 
 interface Props {
     tekster: OppgavebekreftelseTekster;
@@ -20,9 +23,9 @@ interface Props {
 const UløstOppgavebekreftelse = ({ tekster, deltakerNavn, oppgave, children }: Props) => {
     const [visKvittering, setVisKvittering] = useState<boolean>(false);
     const navigate = useNavigate();
+    const { logEvent } = useAnalyticsInstance();
 
     const alertRef = useRef<HTMLDivElement>(null);
-
     const prevVisKvittering = usePrevious(visKvittering);
 
     useEffect(() => {
@@ -31,13 +34,21 @@ const UløstOppgavebekreftelse = ({ tekster, deltakerNavn, oppgave, children }: 
         }
     });
 
+    const handleOnSuccess = (uttalelse: UngdomsytelseOppgaveUttalelseDto) => {
+        setVisKvittering(true);
+        logEvent(
+            LogMetaInfoType.OPPGAVEBEKREFTELSE_SENDT,
+            logUtils.getOppgaveBekreftelseMeta(oppgave, { harUttalelse: uttalelse.harUttalelse }),
+        );
+    };
+
     return (
         <VStack gap="6">
             <div>
                 <OppgaveStatusTag oppgaveStatus={oppgave.status} oppgaveStatusTekst={getOppgaveStatusText(oppgave)} />
             </div>
             <Heading level="1" size="large">
-                {tekster.sidetittel}
+                {tekster.oppgavetittel}
             </Heading>
             {visKvittering ? (
                 <>
@@ -66,7 +77,7 @@ const UløstOppgavebekreftelse = ({ tekster, deltakerNavn, oppgave, children }: 
                     <UtalelseForm
                         harTilbakemeldingSpørsmål={tekster.harTilbakemeldingSpørsmål}
                         oppgaveReferanse={oppgave.oppgaveReferanse}
-                        onSuccess={() => setVisKvittering(true)}
+                        onSuccess={handleOnSuccess}
                         onCancel={() => navigate(AppRoutes.innsyn)}
                     />
                 </>