+++ apps/ungdomsytelse-deltaker/src/apps/innsyn/i18n/messages/nn.ts
@@ -1,3 +1,46 @@
-export const innsynMessages_nn = {
+import { innsynMessages_nb } from './nb';
+
+export const innsynMessages_nn: typeof innsynMessages_nb = {
     'innsyn.sidetittel': 'Din ungdomsprogramytelse',
+
+    'oppgavestatus.LØST': 'Løyst',
+    'oppgavestatus.ULØST': 'Uløyst',
+    'oppgavestatus.AVBRUTT': 'Avbroten',
+    'oppgavestatus.UTLØPT': 'Utgått',
+    'oppgavestatus.LUKKET': 'Lukka',
+
+    'oppgavetype.BEKREFT_ENDRET_STARTDATO.sidetittel': 'Ny startdato i ungdomsprogrammet',
+    'oppgavetype.BEKREFT_ENDRET_STARTDATO.oppgavetittel': 'Sjå og gi tilbakemelding på ny startdato',
+    'oppgavetype.BEKREFT_ENDRET_STARTDATO.info':
+        'Rettleiaren din har endra datoen for når du starta i ungdomsprogrammet.',
+    'oppgavetype.BEKREFT_ENDRET_STARTDATO.harTilbakemeldingSpørsmål': 'Har du ei tilbakemelding på startdatoen?',
+
+    'oppgavetype.BEKREFT_ENDRET_SLUTTDATO.sidetittel': 'Ny sluttdato i ungdomsprogrammet',
+    'oppgavetype.BEKREFT_ENDRET_SLUTTDATO.oppgavetittel': 'Sjå og gi tilbakemelding på sluttdato',
+    'oppgavetype.BEKREFT_ENDRET_SLUTTDATO.info':
+        'Rettleiaren din har sett ein dato for når du sluttar i ungdomsprogrammet.',
+    'oppgavetype.BEKREFT_ENDRET_SLUTTDATO.harTilbakemeldingSpørsmål': 'Har du ei tilbakemelding på sluttdatoen?',
+
+    'oppgavetype.BEKREFT_NY_SLUTTDATO.sidetittel': 'Sluttdato i ungdomsprogrammet',
+    'oppgavetype.BEKREFT_NY_SLUTTDATO.oppgavetittel': 'Sjå og gi tilbakemelding på ny sluttdato',
+    'oppgavetype.BEKREFT_NY_SLUTTDATO.info': 'Rettleiaren din har endra datoen for når du sluttar i ungdomsprogrammet.',
+    'oppgavetype.BEKREFT_NY_SLUTTDATO.harTilbakemeldingSpørsmål': 'Har du ei tilbakemelding på sluttdatoen?',
+
+    'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.sidetittel': 'Korrigert inntekt',
+    'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.oppgavetittel': 'Sjå og gi tilbakemelding på ulik inntekt for {måned}',
+    'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.info':
+        'Det er skilnad mellom det du har oppgjeve i løn, og det arbeidsgjevaren din har rapportert til oss.',
+    'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.harTilbakemeldingSpørsmål':
+        'Forstår og godtek du at inntekta di er endra?',
+
+    'oppgavetype.RAPPORTER_INNTEKT.sidetittel': 'Rapporter inntekt',
+    'oppgavetype.RAPPORTER_INNTEKT.oppgavetittel': 'Meld frå om du hadde inntekt {måned}',
+    'oppgavetype.RAPPORTER_INNTEKT.info':
+        'Meld frå om du jobba og fekk utbetalt løn i september. Hadde du inga inntekt, treng du ikkje gjere noko.',
+    'oppgavetype.RAPPORTER_INNTEKT.info.rapportert': 'Du meldte at du hadde {inntekt} kr i inntekt i {måned}.',
+    'oppgavetype.RAPPORTER_INNTEKT.info.ikkeRapportert': 'Du meldte at du ikkje hadde inntekt i {måned}.',
+
+    'oppgavetype.SØK_YTELSE.sidetittel': 'Søknad for ungdomsprogramytelsen',
+    'oppgavetype.SØK_YTELSE.oppgavetittel': 'Søknad for ungdoms\u00ADprogram\u00ADytelsen',
+    'oppgavetype.SØK_YTELSE.info': 'Du er meldt inn i ungdomsprogrammet. No kan du søkje om ungdomsprogramytelsen.',
 };