+++ apps/ungdomsytelse-deltaker/src/apps/søknad/components/VelkommenMelding.tsx
@@ -1,7 +1,7 @@
 import { BodyLong, Box, GuidePanel, Heading, ReadMore, VStack } from '@navikt/ds-react';
 import { dateFormatter } from '@navikt/sif-common-utils';
 import BehandlingAvPersonopplysningerContent from './BehandlingAvPersonopplysningerContent';
-import { useAppIntl } from '../../../i18n';
+import { AppText, useAppIntl } from '../../../i18n';
 import getLenker from '../../../utils/lenker';
 import ExternalLink from './external-link/ExternalLink';
 
@@ -16,27 +16,34 @@ const VelkommenMelding = ({ fornavn, startdato }: Props) => {
         <GuidePanel poster={true}>
             <Box paddingBlock="4 0">
                 <Heading level="1" size="medium" spacing={true}>
-                    Hei {fornavn}!
+                    <AppText id="velkommenMelding.hei" values={{ fornavn }} />
                 </Heading>
                 <VStack gap="4">
                     <BodyLong>
-                        Du er meldt inn i ungdomsprogrammet fra{' '}
-                        <strong>{dateFormatter.dayDateMonthYear(startdato)}</strong>.
+                        <AppText
+                            id="velkommenMelding.deltakelsePeriode"
+                            values={{
+                                startdato: dateFormatter.dayDateMonthYear(startdato),
+                                strong: (children) => <strong>{children}</strong>,
+                            }}
+                        />
                     </BodyLong>
                     <BodyLong>
-                        Når du er med i ungdomsprogrammet, kan du søke om å få penger. Da får du utbetalt penger hver
-                        måned så lenge du deltar i programmet.
+                        <AppText id="velkommenMelding.ytelseBeskrivelse" />
                     </BodyLong>
                     <BodyLong>
-                        Du søker ved å fylle ut denne søknaden. Les mer om{' '}
-                        <ExternalLink href={getLenker().omUngdomsprogramytelsen}>
-                            ungdomsprogrammet og -ytelsen på nav.no
-                        </ExternalLink>
-                        .
+                        <AppText
+                            id="velkommenMelding.søknadBeskrivelse"
+                            values={{
+                                Lenke: (children) => (
+                                    <ExternalLink href={getLenker().omUngdomsprogramytelsen}>{children}</ExternalLink>
+                                ),
+                            }}
+                        />
                     </BodyLong>
                     <VStack gap="0">
-                        <ReadMore header="Hva hvis datoen ikke stemmer?">
-                            Da tar du kontakt med din veileder og sier fra om dette, før du sender inn denne søknaden.
+                        <ReadMore header={text('velkommenMelding.readMore.dato.header')}>
+                            <AppText id="velkommenMelding.readMore.dato.content" />
                         </ReadMore>
                         <ReadMore header={text('personopplysninger.accordion.header')}>
                             <BehandlingAvPersonopplysningerContent />