+++ apps/ungdomsytelse-deltaker/src/apps/søknad/context/SøknadContext.tsx
@@ -4,6 +4,7 @@ import { RegistrertBarn, Søker } from '@navikt/sif-common-api';
 import { YesOrNo } from '@navikt/sif-common-core-ds/src';
 import { DeltakelsePeriode, formaterKontonummer, SøkYtelseOppgave } from '@navikt/ung-common';
 import { ApplikasjonHendelse, useAnalyticsInstance } from '../../../utils/analytics';
+import { LogMetaInfoType, logUtils } from '../../innsyn/utils/logUtils';
 import { MellomlagringDTO } from '../api/mellomlagring/mellomlagring';
 import { useSøknadNavigation } from '../hooks/utils/useSøknadNavigation';
 import { Spørsmål, Steg, SøknadContextType, SøknadSvar } from '../types';
@@ -34,7 +35,7 @@ export const SøknadProvider = ({
 }: SøknadProviderProps) => {
     const [svar, setSvar] = useState<SøknadSvar>(initialSvar || initialData);
     const { gotoSteg, gotoVelkommenPage } = useSøknadNavigation();
-    const { logHendelse, logInfo, logSoknadStartet, logSoknadSent } = useAnalyticsInstance();
+    const { logHendelse, logEvent, logSoknadStartet, logSoknadSent } = useAnalyticsInstance();
 
     const [søknadSendt, setSøknadSendt] = useState(false);
     const [søknadStartet, setSøknadStartet] = useState(initialData.harForståttRettigheterOgPlikter ? true : false);
@@ -62,12 +63,15 @@ export const SøknadProvider = ({
     const doSetSøknadSendt = () => {
         setSøknadSendt(true);
         logSoknadSent(UngdomsytelseDeltakerApp.key);
-        logInfo({
-            harBarn: barn.length > 0,
-            barnStemmer: svar[Spørsmål.BARN] === YesOrNo.YES,
-            harKontonummer: kontonummer ? true : false,
-            kontonummerStemmer: svar[Spørsmål.KONTONUMMER] === YesOrNo.YES,
-        });
+        logEvent(
+            LogMetaInfoType.SØKNAD_SENDT,
+            logUtils.getSøknadInnsendingMeta(deltakelsePeriode, {
+                antallBarn: barn.length,
+                barnStemmer: svar[Spørsmål.BARN] === YesOrNo.YES,
+                harKontonummer: kontonummer ? true : false,
+                kontonummerStemmer: svar[Spørsmål.KONTONUMMER] === YesOrNo.YES,
+            }),
+        );
     };
 
     return (