+++ apps/ungdomsytelse-deltaker/src/apps/søknad/i18n/messages/nb.ts
@@ -1,6 +1,7 @@
 /* eslint-disable max-len */
 export const ungSoknadMessages_nb = {
-    'soknad.title': 'Søknad om ungdomsprogramytelsen',
+    'søknad.tittel': 'Søknad om ungdomsprogramytelsen',
+    'søknad.tittel.shy': 'Søknad om ungdoms\u00ADprogram\u00ADytelsen',
 
     'personopplysninger.accordion.header': 'Om hvordan vi innhenter opplysninger om deg',
     'personopplysninger.1': 'Slik behandler Nav personopplysningene dine',
@@ -19,4 +20,94 @@ export const ungSoknadMessages_nb = {
         'opplysninger om perioden du deltar i programmet, henter vi fra veileder i Nav, slik at vi vet for hvilken periode du skal motta ytelse/penger fra Nav',
     'personopplysninger.5':
         'Du har rett til innsyn i saken din. Vil du vite mer om hvordan Nav behandler personopplysninger? Se <Lenke>nav.no/personvern</Lenke>.',
+
+    'søknadApp.loading.error': 'Det oppstod en feil ved henting av nødvendig informasjon. Vennligst prøv igjen senere.',
+
+    'søknadApp.nesteSteg.label': 'Neste steg',
+    'søknadApp.forrigeSteg.label': 'Forrige steg',
+    'søknadApp.avbrytSøknad.label': 'Avbryt søknad',
+    'søknadApp.sendSøknad.label': 'Send søknad',
+
+    'søknadApp.slettSøknad.tittel': 'Slett søknad',
+    'søknadApp.slettSøknad.ja.label': 'Ja, slett',
+    'søknadApp.slettSøknad.nei.label': 'Nei',
+    'søknadApp.slettSøknad.spm': 'Ønsker du å slette søknaden?',
+    'søknadApp.slettSøknad.tekst': 'Informasjonen du har fylt ut blir slettet, og du kommer tilbake til velkomstsiden.',
+
+    /** VelkommenPage */
+    'velkommenPage.startSøknad': 'Start søknad',
+    'velkommenPage.bekrefter': 'Jeg vil svare så godt jeg kan på spørsmålene i søknaden.',
+    'velkommenPage.infoStemmer':
+        'Det er viktig at du gir oss riktige opplysninger slik at vi kan behandle saken din. <Lenke>Les mer om viktigheten av å gi riktige opplysninger</Lenke>.',
+    'velkommenPage.bekreftelse.skjultLegend': 'Bekreftelse',
+    'velkommenPage.validering.bekrefterIkkeValgt': 'Du må bekrefte at du vil svare så riktig som du kan.',
+
+    /** VelkommenMelding */
+    'velkommenMelding.hei': 'Hei {fornavn}!',
+    'velkommenMelding.deltakelsePeriode': 'Du er meldt inn i ungdomsprogrammet fra <strong>{startdato}</strong>.',
+    'velkommenMelding.ytelseBeskrivelse':
+        'Når du er med i ungdomsprogrammet, kan du søke om å få penger. Da får du utbetalt penger hver måned så lenge du deltar i programmet.',
+    'velkommenMelding.søknadBeskrivelse':
+        'Du søker ved å fylle ut denne søknaden. Les mer om <Lenke>ungdomsprogrammet og -ytelsen på nav.no</Lenke>.',
+    'velkommenMelding.readMore.dato.header': 'Hva hvis datoen ikke stemmer?',
+    'velkommenMelding.readMore.dato.content':
+        'Da tar du kontakt med din veileder og sier fra om dette, før du sender inn denne søknaden.',
+
+    /** KvitteringPage */
+    'kvitteringPage.dokumentTittel': 'Søknad er sendt - Søknad om ungdomsprogramytelsen',
+    'kvitteringPage.tittel': 'Søknaden er sendt!',
+    'kvitteringPage.beskrivelse': 'Vi har fått søknaden din om penger gjennom ungdomsprogramytelsen.',
+    'kvitteringPage.hvaSkjerVidere': 'Hva skjer videre?',
+    'kvitteringPage.hvaSkjerVidere.1':
+        'Du får en SMS når søknaden din er ferdig behandlet. Da kan du logge inn på Min side på nav.no og se svaret på søknaden (vedtaksbrevet).',
+    'kvitteringPage.hvaSkjerVidere.2': 'Pengene får du som regel den 10. hver måned.',
+    'kvitteringPage.hvaSkjerVidere.3':
+        'Hvis du tjener penger mens du er i ungdomsprogrammet, må du melde fra om dette. Du får en SMS den 1. hver måned, og så logger du inn på Min side på nav.no og melder fra om hva du tjente måneden før.',
+    'kvitteringPage.lykkeTil': 'Vi ønsker deg lykke til med veien videre!',
+    'kvitteringPage.gåTilMinSide': 'Gå til Min side',
+
+    /** KontonummerSteg */
+    'kontonummerSteg.tittel': 'Kontonummer for utbetaling',
+    'kontonummerSteg.beskrivelse':
+        'For å få pengene inn på bankkontoen din, må du ha registrert kontonummeret ditt hos Nav.',
+    'kontonummerSteg.kontonummer.spm': 'Er kontonummeret ditt {kontonummer}?',
+    'kontonummerSteg.kontonummer.ja.label': 'Ja',
+    'kontonummerSteg.kontonummer.nei.label': 'Nei',
+    'kontonummerSteg.kontonummer.stemmerIkke.info':
+        'Gå til <Lenke>personopplysninger på Min side</Lenke> for å endre kontonummeret ditt.',
+    'kontonummerSteg.kontonummer.stemmerIkke.info.2':
+        'Vi anbefaler at du endrer kontonummeret ditt før du sender inn søknaden, slik at pengene blir satt inn på kontoen din.',
+    'kontonummerSteg.harIkkeKontonummer.info.1': 'Du har ikke registrert kontonummer hos oss',
+    'kontonummerSteg.harIkkeKontonummer.info.2':
+        'Registrer bankkontonummeret ditt hos Nav slik at du får pengene utbetalt til rett konto. Gå til <Lenke>personopplysninger på Min side</Lenke> for å legge inn kontonummeret ditt.',
+    'kontonummerSteg.harIkkeKontonummer.info.3':
+        'Du kan fremdeles sende inn søknaden, men vi anbefaler at du legger inn kontonummeret med én gang slik at pengene ikke blir forsinket.',
+
+    /** BarnSteg */
+    'barnSteg.tittel': 'Barn',
+    'barnSteg.beskrivelse': 'Hvis du deltar i ungdomsprogrammet og har barn, har du rett på et barnetillegg.',
+    'barnSteg.registrerteBarn.tittel': 'Barn vi har registrert på deg:',
+    'barnSteg.barnStemmer.ja.label': 'Ja',
+    'barnSteg.barnStemmer.nei.label': 'Nei',
+    'barnSteg.spørsmål.ingenBarn': 'Stemmer det at du ikke har barn?',
+    'barnSteg.spørsmål.harBarn': 'Stemmer informasjonen om {antallBarn, plural, one {barnet} other {barna}}?',
+    'barnSteg.validering.ikkeSvart': 'Du må svare på om informasjonen stemmer',
+    'barnSteg.opplysninger.info.tittel': 'Vi henter opplysninger fra folkeregisteret',
+    'barnSteg.opplysninger.info.text':
+        'Du må være registrert som forelder med foreldreansvar i Folkeregisteret for å ha rett på barnetillegg. Hvis du mener opplysningene fra Folkeregisteret er feil, må du ta <Lenke>kontakt med Skatteetaten</Lenke>. Hos Skatteetaten kan du registrere foreldreansvar.',
+    'barnSteg.barnInfo.ingenBarn': 'Vi har ikke registrert at du har barn.',
+
+    /** OppsummeringSteg */
+    'oppsummeringSteg.tittel': 'Oppsummering',
+    'oppsummeringSteg.startdato': 'Startdato',
+    'oppsummeringSteg.kontonummer.tittel': 'Kontonummer for utbetaling',
+    'oppsummeringSteg.kontonummer.ingenKontonummer.tittel': 'Kontonummer for utbetaling',
+    'oppsummeringSteg.kontonummer.ingenKontonummer.tekst': 'Vi har ikke registrert noe kontonummer på deg.',
+    'oppsummeringSteg.barn.tittel': 'Barn',
+    'oppsummeringSteg.bekreft.hiddenLegend': 'Bekreft opplysninger',
+    'oppsummeringSteg.bekreft.tekst':
+        'Jeg bekrefter at opplysningene over er riktige og at jeg ønsker søke om ungdoms\u00ADprogram\u00ADytelsen.',
+    'oppsummeringSteg.bekreft.validering.bekreftIkkeValgt': 'Du må bekrefte at opplysningene er riktige',
+    'oppsummeringSteg.søknadIkkeGyldig':
+        'Det mangler opplysninger i søknaden din. Vennligst gå tilbake og fyll ut de nødvendige feltene.',
 };