+++ apps/ungdomsytelse-deltaker/src/apps/søknad/steg/barn/BarnInfo.tsx
@@ -1,12 +1,19 @@
 import { BodyShort } from '@navikt/ds-react';
 import { RegistrertBarn } from '@navikt/sif-common-api';
 import BarnList from './BarnList';
+import { AppText } from '../../../../i18n';
 
 interface Props {
     barn: RegistrertBarn[];
 }
 
 const BarnInfo = ({ barn }: Props) =>
-    barn.length > 0 ? <BarnList barn={barn} /> : <BodyShort>Vi har ikke registrert at du har barn.</BodyShort>;
+    barn.length > 0 ? (
+        <BarnList barn={barn} />
+    ) : (
+        <BodyShort>
+            <AppText id="barnSteg.barnInfo.ingenBarn" />
+        </BodyShort>
+    );
 
 export default BarnInfo;