+++ apps/ungdomsytelse-deltaker/src/apps/søknad/steg/oppsummering/OppsummeringSteg.tsx
@@ -1,7 +1,9 @@
 import { Alert, Checkbox, CheckboxGroup, FormSummary, VStack } from '@navikt/ds-react';
 import { useState } from 'react';
 import { YesOrNo } from '@navikt/sif-common-core-ds/src';
+import { dateFormatter } from '@navikt/sif-common-utils';
 import ApiErrorAlert from '@navikt/ung-common/src/components/api-error-alert/ApiErrorAlert';
+import { AppText, useAppIntl } from '../../../../i18n';
 import SkjemaFooter from '../../components/steg-skjema/SkjemaFooter';
 import SøknadSteg from '../../components/søknad-steg/SøknadSteg';
 import { useSendSøknad } from '../../hooks/api/useSendSøknad';
@@ -9,13 +11,15 @@ import { useSøknadContext } from '../../hooks/context/useSøknadContext';
 import { useSøknadNavigation } from '../../hooks/utils/useSøknadNavigation';
 import { Spørsmål, Steg } from '../../types';
 import BarnInfo from '../barn/BarnInfo';
-import { getBarnSpørsmål } from '../barn/BarnSteg';
 import { buildSøknadFromSvar } from './oppsummeringUtils';
-import { dateFormatter } from '@navikt/sif-common-utils';
+import { useAnalyticsInstance } from '../../../../utils/analytics';
+import { UngdomsytelseDeltakerApp } from '@navikt/sif-app-register';
 
 const OppsummeringSteg = () => {
+    const { text } = useAppIntl();
     const { søker, deltakelsePeriode, søknadOppgave, setSøknadSendt, kontonummerInfo, barn, svar } = useSøknadContext();
     const { gotoSteg, gotoKvittering } = useSøknadNavigation();
+    const { logSoknadFailed } = useAnalyticsInstance();
 
     const [bekrefterOpplysninger, setBekrefterOpplysninger] = useState<boolean>(false);
     const [bekreftError, setBekreftError] = useState<string | undefined>();
@@ -30,15 +34,13 @@ const OppsummeringSteg = () => {
         kontonummerInfo.harKontonummer ? kontonummerInfo.kontonummerFraRegister : undefined,
     );
 
-    const søknadError = søknad
-        ? undefined
-        : 'Det mangler opplysninger i søknaden din. Vennligst gå tilbake og fyll ut de nødvendige feltene.';
+    const søknadError = søknad ? undefined : text('oppsummeringSteg.søknadIkkeGyldig');
 
     const handleOnSubmit = async () => {
         if (søknad) {
             setBekreftError(undefined);
             if (!bekrefterOpplysninger) {
-                setBekreftError('Du må bekrefte at opplysningene er riktige');
+                setBekreftError(text('oppsummeringSteg.bekreft.validering.bekreftIkkeValgt'));
                 return;
             }
             try {
@@ -46,22 +48,27 @@ const OppsummeringSteg = () => {
                 setSøknadSendt();
                 gotoKvittering();
             } catch {
-                // Håndteres gjennom error objektet  i useSendSøknad
+                // Håndteres gjennom error objektet i useSendSøknad
+                logSoknadFailed(UngdomsytelseDeltakerApp.key);
             }
         }
     };
 
     return (
-        <SøknadSteg tittel="Oppsummering" steg={Steg.OPPSUMMERING}>
+        <SøknadSteg tittel={text('oppsummeringSteg.tittel')} steg={Steg.OPPSUMMERING}>
             <VStack gap="6">
                 <VStack gap="4">
                     <FormSummary>
                         <FormSummary.Header>
-                            <FormSummary.Heading level="2">Ungdomsprogramytelsen</FormSummary.Heading>
+                            <FormSummary.Heading level="2">
+                                <AppText id="søknad.tittel" />
+                            </FormSummary.Heading>
                         </FormSummary.Header>
                         <FormSummary.Answers>
                             <FormSummary.Answer>
-                                <FormSummary.Label>Startdato</FormSummary.Label>
+                                <FormSummary.Label>
+                                    <AppText id="oppsummeringSteg.startdato" />
+                                </FormSummary.Label>
                                 <FormSummary.Value>
                                     {dateFormatter.compact(deltakelsePeriode.programPeriode.from)}
                                 </FormSummary.Value>
@@ -70,7 +77,9 @@ const OppsummeringSteg = () => {
                     </FormSummary>
                     <FormSummary>
                         <FormSummary.Header>
-                            <FormSummary.Heading level="2">Kontonummer for utbetaling</FormSummary.Heading>
+                            <FormSummary.Heading level="2">
+                                <AppText id="oppsummeringSteg.kontonummer.tittel" />
+                            </FormSummary.Heading>
                             <FormSummary.EditLink
                                 href="#"
                                 onClick={(evt) => {
@@ -84,17 +93,24 @@ const OppsummeringSteg = () => {
                             {kontonummerInfo.harKontonummer ? (
                                 <FormSummary.Answer>
                                     <FormSummary.Label>
-                                        Stemmer det at kontonummeret ditt er {kontonummerInfo.formatertKontonummer}?
+                                        <AppText
+                                            id="kontonummerSteg.kontonummer.spm"
+                                            values={{ kontonummer: kontonummerInfo.formatertKontonummer }}
+                                        />
                                     </FormSummary.Label>
                                     <FormSummary.Value>
-                                        {svar[Spørsmål.KONTONUMMER] === YesOrNo.YES ? 'Ja' : 'Nei'}
+                                        {svar[Spørsmål.KONTONUMMER] === YesOrNo.YES
+                                            ? text('kontonummerSteg.kontonummer.ja.label')
+                                            : text('kontonummerSteg.kontonummer.nei.label')}
                                     </FormSummary.Value>
                                 </FormSummary.Answer>
                             ) : (
                                 <FormSummary.Answer>
-                                    <FormSummary.Label>Kontonummer for utbetaling</FormSummary.Label>
+                                    <FormSummary.Label>
+                                        <AppText id="oppsummeringSteg.kontonummer.ingenKontonummer.tittel" />
+                                    </FormSummary.Label>
                                     <FormSummary.Value>
-                                        Vi har ikke registrert noe kontonummer på deg.
+                                        <AppText id="oppsummeringSteg.kontonummer.ingenKontonummer.tekst" />
                                     </FormSummary.Value>
                                 </FormSummary.Answer>
                             )}
@@ -102,7 +118,9 @@ const OppsummeringSteg = () => {
                     </FormSummary>
                     <FormSummary>
                         <FormSummary.Header>
-                            <FormSummary.Heading level="2">Barn</FormSummary.Heading>
+                            <FormSummary.Heading level="2">
+                                <AppText id="oppsummeringSteg.barn.tittel" />
+                            </FormSummary.Heading>
                             <FormSummary.EditLink
                                 href="#"
                                 onClick={(evt) => {
@@ -114,15 +132,30 @@ const OppsummeringSteg = () => {
                         </FormSummary.Header>
                         <FormSummary.Answers>
                             <FormSummary.Answer>
-                                <FormSummary.Label>Barn vi har registrert på deg:</FormSummary.Label>
+                                <FormSummary.Label>
+                                    <AppText id="barnSteg.registrerteBarn.tittel" />
+                                </FormSummary.Label>
                                 <FormSummary.Value>
                                     <BarnInfo barn={barn} />
                                 </FormSummary.Value>
                             </FormSummary.Answer>
                             <FormSummary.Answer>
-                                <FormSummary.Label>{getBarnSpørsmål(barn.length)}</FormSummary.Label>
+                                <FormSummary.Label>
+                                    <AppText
+                                        id={
+                                            barn.length === 0
+                                                ? 'barnSteg.spørsmål.ingenBarn'
+                                                : 'barnSteg.spørsmål.harBarn'
+                                        }
+                                        values={{
+                                            antallBarn: barn.length,
+                                        }}
+                                    />
+                                </FormSummary.Label>
                                 <FormSummary.Value>
-                                    {svar[Spørsmål.BARN] === YesOrNo.YES ? 'Ja' : 'Nei'}
+                                    {svar[Spørsmål.BARN] === YesOrNo.YES
+                                        ? text('barnSteg.barnStemmer.ja.label')
+                                        : text('barnSteg.barnStemmer.nei.label')}
                                 </FormSummary.Value>
                             </FormSummary.Answer>
                         </FormSummary.Answers>
@@ -141,7 +174,7 @@ const OppsummeringSteg = () => {
                         <CheckboxGroup
                             name="bekrefterInnsending"
                             hideLegend={true}
-                            legend="Bekreft opplysninger"
+                            legend={text('oppsummeringSteg.bekreft.hiddenLegend')}
                             error={bekreftError}>
                             <Checkbox
                                 value="bekrefter"
@@ -149,16 +182,22 @@ const OppsummeringSteg = () => {
                                     setBekreftError(undefined);
                                     setBekrefterOpplysninger(evt.target.checked);
                                 }}>
-                                Jeg bekrefter at opplysningene over er riktige og at jeg ønsker søke om
-                                ungdomsprogram&shy;ytelsen.
+                                <AppText id="oppsummeringSteg.bekreft.tekst" />
                             </Checkbox>
                         </CheckboxGroup>
 
                         {error ? <ApiErrorAlert error={error} /> : null}
                         <SkjemaFooter
                             pending={isPending}
-                            forrige={{ tittel: 'Forrige steg', onClick: () => gotoSteg(Steg.BARN) }}
-                            submit={{ tittel: 'Send søknad', disabled: !!søknadError, erSendInn: true }}
+                            forrige={{
+                                tittel: text('søknadApp.forrigeSteg.label'),
+                                onClick: () => gotoSteg(Steg.BARN),
+                            }}
+                            submit={{
+                                tittel: text('søknadApp.sendSøknad.label'),
+                                disabled: !!søknadError,
+                                erSendInn: true,
+                            }}
                         />
                     </VStack>
                 </form>